Refactor of pages
