+++ src/about.js
@@ -1,0 +1,13 @@
+
+import React from 'react'
+import { Row } from 'react-bootstrap'
+
+function About() {
+    return (
+        <Row>
+          Basic Application for data entry
+        </Row>
+    )
+}
+
+export { About }
