+++ src/dashboard.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Row } from 'react-bootstrap'
+
+function Dashboard() {
+    return (
+      <Row>Dashboard</Row>
+    )
+}
+
+export { Dashboard }
